refactor: moved layouts inside the components folder
- Layouts belong as components.
