+++ src/components/layouts/MainLayout.tsx
@@ -1,9 +1,9 @@
-import Header from "../components/header/Header";
-import Footer from "../components/footer/Footer";
+import Header from "../header/Header";
+import Footer from "../footer/Footer";
 
 import type { ReactNode } from "react";
-import type { NavigationNode } from "../lib/build-site-navigation";
-import { Toaster } from "../components/toast/toaster";
+import type { NavigationNode } from "../../lib/build-site-navigation";
+import { Toaster } from "../toast/toaster";
 
 interface IMainLayout {
   nav?: NavigationNode[];
